PhotoViewModel was seperated to PhotoCellViewModel and PhotoDetailViewModel.
